backend/pyproject.toml
[project]
name = "gestion-financiere"
version = "1.0.0"
description = "Application de gestion financiere personnelle"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "numpy",
    "plotly",
    "matplotlib",
    "Pillow",
    "opencv-python",
    "pyyaml",
    "pdfminer.six",
    "pdfplumber",
    "rapidocr-onnxruntime",
    "python-dateutil",
    "psutil",
    "pydantic",
    "requests",
    "pyjwt",
    "cryptography",
    "python-dotenv",
    "ollama>=0.6.1",
    "groq>=1.0.0",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
    "python-multipart>=0.0.22",
    "pybrowsers>=1.3.2",
    "platformdirs>=4.0.0",
    "sqlcipher3",
]

[project.scripts]

[project.optional-dependencies]
dev = []
test = [
    "pytest",
]

[dependency-groups]
build = [
    "pyinstaller>=6.0",  # Windows uniquement (PyInstaller + Inno Setup)
]
dev = [
    "pandas-stubs~=2.3.3",
]
test = [
    "pytest>=9.0.2",
]

[tool.setuptools.packages.find]
include = ["domains*", "shared*", "config*", "resources*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.css", "*.png", "*.ico"]
